refactor: replace yup schema into separate file
